requirements.txt
# Core dependencies - required for basic functionality
numpy>=1.21.0
scipy>=1.7.0
pyscf>=2.1.0
ase>=3.22.0
mdanalysis>=2.4.0
biopython>=1.80
numba>=0.56.0
matplotlib>=3.5.0
plotly>=5.0.0
py3dmol>=2.0.0
nglview>=3.0.0
pandas>=1.4.0
h5py>=3.6.0
scikit-image>=0.18.0
tqdm>=4.64.0
pyyaml>=6.0
requests>=2.25.0

# GPU acceleration dependencies - required
cupy-cuda11x>=11.0.0

# Optional dependencies (install as needed)
# Parallel computing: mpi4py>=3.1.0
# Development tools: pytest>=7.0.0 black>=22.0.0 flake8>=4.0.0 mypy>=0.950

# Note: If you have CUDA 12.x instead of 11.x, replace cupy-cuda11x with cupy-cuda12x